pipelines/ingest_tmdb/pyproject.toml
[project]
name = "tmdb-pipelines"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
  "pandas>=2.2",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.1",
  "python-dotenv>=1.0",
  "httpx>=0.27",
  "pgvector>=0.3",
]
